MULE-18516: TCP Connector is only reading the first packet of a multipacket message, truncating the payload when using xml-protocol
There was a criteria for deciding whether to continue reading from the input stream or not that relied on the read method using the whole buffer to the last byte. That is not guaranteed by the read method of InputStream. If the whole buffer was not used there might still be data left to read.

Fixed the isRepeat method and the assertions in the test that reproduced the behaviour. If the input stream is "slow" it should still read the data completely.